Further optimization and repairing for ui-tests
- tuning some timing and some testing strategy
- add focus to input and textarea result sections in order to improve demo experience; if the demo is viewed and used embedded within the documentation page, the user will not jump to some arbitrary hyperlink down on the page, but stay within the demo and view the updated result.

Running the tests by `npx playwright test --project=chromium` produces a stable green experience on my computer. Sometimes some of the more complex tests (datacollection ones) may be flaky - but this appears seldom and seem to be directly linked to heavy load on the machine.
