+++ headless-demo/playwright.config.ts
@@ -36,7 +36,7 @@ const config: PlaywrightTestConfig = {
     /* Base URL to use in actions like `await page.goto('/')`. */
     //baseURL: 'https://next.fritz2.dev/headless-demo/',
     //baseURL: 'C:\\Users\\DE-P0029\\IdeaProjects\\fritz2\\headless-demo\\build\\distributions\\index.html',
-    baseURL: 'http://localhost:8080/',
+    baseURL: 'http://localhost:8081/',
 
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
